requirements.txt
customtkinter==5.1.3
Pillow==9.4.0
Pillow==10.0.0
